Move fixed stack to a separate file
